requirements.txt
flask
scikit-learn
networkx
bcrypt
PyJWT
Flask-SQLAlchemy
gunicorn
dotenv